rebased and updated changes
