+++ workspaces/tekton/package.json
@@ -8,8 +8,6 @@
   "scripts": {
     "dev": "yarn start",
     "start": "backstage-cli repo start",
-    "start:alpha": "yarn workspaces foreach -A --include @backstage-community/plugin-tekton --include backend --parallel -v -i run start:alpha",
-    "start-backend": "yarn workspace backend start",
     "tsc": "tsc",
     "tsc:full": "tsc --skipLibCheck true --incremental false",
     "build:all": "backstage-cli repo build --all",
@@ -19,10 +17,10 @@
     "clean": "backstage-cli repo clean",
     "test": "backstage-cli repo test",
     "test:all": "backstage-cli repo test --coverage",
+    "test:nfs": "playwright test",
     "test:legacy": "APP_MODE=legacy playwright test",
-    "test:alpha": "APP_MODE=alpha playwright test",
     "test:e2e": "playwright test",
-    "test:e2e:ci": "yarn test:legacy && yarn test:alpha",
+    "test:e2e:ci": "yarn test:nfs && yarn test:legacy",
     "playwright": "sh -c 'if [ \"$1\" = test ] && [ $# -eq 1 ]; then yarn test:e2e:ci; else exec playwright \"$@\"; fi' _",
     "fix": "backstage-cli repo fix",
     "lint": "backstage-cli repo lint --since origin/main",
