+++ workspaces/tekton/playwright.config.ts
@@ -16,10 +16,10 @@
 
 import { defineConfig } from '@playwright/test';
 
-// APP_MODE: 'legacy' (dev/index.mock.tsx) or 'alpha' (dev/alpha/index.mock.tsx)
-const appMode = process.env.APP_MODE || 'legacy';
+// APP_MODE: 'legacy' (dev/legacy.mock.tsx) or default NFS (dev/index.mock.tsx)
+const appMode = process.env.APP_MODE || 'nfs';
 const startCommand =
-  appMode === 'legacy' ? 'yarn start:mock' : 'yarn start:alpha:mock';
+  appMode === 'legacy' ? 'yarn start:legacy:mock' : 'yarn start:mock';
 
 /**
  * See https://playwright.dev/docs/test-configuration.
