+++ workspaces/tekton/plugins/tekton/dev/index.mock.tsx
@@ -14,54 +14,157 @@
  * limitations under the License.
  */
 
-import { Page, Header, TabbedLayout } from '@backstage/core-components';
-import { createDevApp } from '@backstage/dev-utils';
-import { EntityProvider } from '@backstage/plugin-catalog-react';
+/**
+ * Playwright NFS entry. Same mocked NFS app as `yarn start` (`dev/index.tsx`).
+ *
+ * Must be self-contained: a static `import './index'` compiles under rspack
+ * but React never mounts, so Playwright sees a blank page.
+ */
+
+import '@backstage/cli/asset-types';
+// eslint-disable-next-line @backstage/no-ui-css-imports-in-non-frontend
+import '@backstage/ui/css/styles.css';
+
+import ReactDOM from 'react-dom/client';
+
+import { createApp } from '@backstage/frontend-defaults';
+import { SignInPage } from '@backstage/core-components';
+import {
+  ApiBlueprint,
+  createFrontendModule,
+  pluginHeaderActionsApiRef,
+} from '@backstage/frontend-plugin-api';
+import { SignInPageBlueprint } from '@backstage/plugin-app-react';
+import { catalogApiRef } from '@backstage/plugin-catalog-react';
+import catalogPlugin from '@backstage/plugin-catalog/alpha';
+import kubernetesPlugin from '@backstage/plugin-kubernetes/alpha';
 import {
   kubernetesApiRef,
-  kubernetesProxyApiRef,
   kubernetesAuthProvidersApiRef,
+  kubernetesProxyApiRef,
 } from '@backstage/plugin-kubernetes-react';
+import { AuthorizeResult } from '@backstage/plugin-permission-common';
 import { permissionApiRef } from '@backstage/plugin-permission-react';
-import { TestApiProvider } from '@backstage/test-utils';
-import { tektonTranslations } from '@backstage-community/plugin-tekton';
 
-import { TektonCI, tektonPlugin } from '../src/plugin';
+import tektonPlugin from '../src';
+import tektonTranslationsModule from '../src/translations';
+import { devSidebarContent } from './shared';
 import {
-  mockEntity,
+  mockCatalogApi,
   mockKubernetesAuthProviderApi,
   mockKubernetesClient,
   mockKubernetesProxyApi,
-  mockPermissionApi,
 } from './mocks';
 
-createDevApp()
-  .addTranslationResource(tektonTranslations)
-  .setAvailableLanguages(['en', 'de', 'es', 'fr', 'it', 'ja'])
-  .addPage({
-    element: (
-      <TestApiProvider
-        apis={[
-          [kubernetesApiRef, mockKubernetesClient],
-          [kubernetesProxyApiRef, mockKubernetesProxyApi],
-          [permissionApiRef, mockPermissionApi],
-          [kubernetesAuthProvidersApiRef, mockKubernetesAuthProviderApi],
-        ]}
-      >
-        <EntityProvider entity={mockEntity}>
-          <Page themeId="service">
-            <Header type="component — service" title="demo-sevice" />
-            <TabbedLayout>
-              <TabbedLayout.Route path="/" title="CI/CD">
-                <TektonCI />
-              </TabbedLayout.Route>
-            </TabbedLayout>
-          </Page>
-        </EntityProvider>
-      </TestApiProvider>
-    ),
-    title: 'Tekton CI',
-    path: '/tekton',
-  })
-  .registerPlugin(tektonPlugin)
-  .render();
+const catalogPluginOverrides = catalogPlugin.withOverrides({
+  extensions: [
+    catalogPlugin.getExtension('api:catalog').override({
+      params: defineParams =>
+        defineParams({
+          api: catalogApiRef,
+          deps: {},
+          factory: () => mockCatalogApi,
+        }),
+    }),
+  ],
+});
+
+const kubernetesPluginOverrides = kubernetesPlugin.withOverrides({
+  extensions: [
+    kubernetesPlugin.getExtension('api:kubernetes').override({
+      params: defineParams =>
+        defineParams({
+          api: kubernetesApiRef,
+          deps: {},
+          factory: () => mockKubernetesClient,
+        }),
+    }),
+    kubernetesPlugin.getExtension('api:kubernetes/proxy').override({
+      params: defineParams =>
+        defineParams({
+          api: kubernetesProxyApiRef,
+          deps: {},
+          factory: () => mockKubernetesProxyApi,
+        }),
+    }),
+    kubernetesPlugin.getExtension('api:kubernetes/auth-providers').override({
+      params: defineParams =>
+        defineParams({
+          api: kubernetesAuthProvidersApiRef,
+          deps: {},
+          factory: () => mockKubernetesAuthProviderApi,
+        }),
+    }),
+  ],
+});
+
+const signInPage = SignInPageBlueprint.make({
+  params: {
+    loader: async () => props =>
+      (
+        <SignInPage
+          {...props}
+          title="Select a sign-in method"
+          align="center"
+          providers={['guest']}
+        />
+      ),
+  },
+});
+
+const appDevModule = createFrontendModule({
+  pluginId: 'app',
+  extensions: [
+    signInPage,
+    ApiBlueprint.make({
+      name: 'permission',
+      params: defineParams =>
+        defineParams({
+          api: permissionApiRef,
+          deps: {},
+          factory: () => ({
+            authorize: async () => ({
+              result: window.location.pathname.includes('permission-denied')
+                ? AuthorizeResult.DENY
+                : AuthorizeResult.ALLOW,
+            }),
+          }),
+        }),
+    }),
+    ApiBlueprint.make({
+      name: 'plugin-header-actions',
+      params: defineParams =>
+        defineParams({
+          api: pluginHeaderActionsApiRef,
+          deps: {},
+          factory: () => ({
+            getPluginHeaderActions: () => [],
+          }),
+        }),
+    }),
+  ],
+});
+
+const devNavModule = createFrontendModule({
+  pluginId: 'app',
+  extensions: [devSidebarContent],
+});
+
+const app = createApp({
+  features: [
+    catalogPluginOverrides,
+    kubernetesPluginOverrides,
+    tektonPlugin,
+    tektonTranslationsModule,
+    devNavModule,
+    appDevModule,
+  ],
+});
+
+if (window.location.pathname === '/') {
+  window.location.replace('/catalog');
+}
+
+const root = app.createRoot();
+
+ReactDOM.createRoot(document.getElementById('root')!).render(root);
