+++ workspaces/tekton/plugins/tekton/report-alpha.api.md
@@ -4,7 +4,7 @@
 
 ```ts
 import { FrontendModule } from '@backstage/frontend-plugin-api';
-import { TranslationRef } from '@backstage/core-plugin-api/alpha';
+import { TranslationRef } from '@backstage/frontend-plugin-api';
 import { TranslationResource } from '@backstage/frontend-plugin-api';
 
 // @public
