+++ workspaces/tekton/plugins/tekton/src/plugin.test.ts
@@ -18,5 +18,6 @@ import tektonPlugin from './plugin';
 describe('tekton', () => {
   it('should export plugin', () => {
     expect(tektonPlugin).toBeDefined();
+    expect(tektonPlugin.pluginId).toBe('tekton');
   });
 });
