+++ workspaces/tekton/plugins/tekton/src/plugin.tsx
@@ -13,10 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-import {
-  createFrontendPlugin,
-  type FrontendPlugin,
-} from '@backstage/frontend-plugin-api';
+import { createFrontendPlugin } from '@backstage/frontend-plugin-api';
 import { EntityContentBlueprint } from '@backstage/plugin-catalog-react/alpha';
 import { isTektonCIAvailable } from './utils/isTektonCIAvailable';
